@@ -11,6 +11,7 @@
     "@types/react": "^18.2.6",
     "@types/react-dom": "^18.2.4",
     "react": "^18.2.0",
+    "uuid": "^9.0.1",
     "web-vitals": "^2.1.4"
   },
   "scripts": {
@@ -38,26 +39,37 @@
     ]
   },
   "devDependencies": {
-    "@babel/core": "^7.21.8",
-    "@babel/preset-env": "^7.21.5",
-    "@babel/preset-react": "^7.18.6",
-    "@babel/preset-typescript": "^7.21.5",
-    "@typescript-eslint/eslint-plugin": "^5.59.7",
+    "@babel/core": "^7.22.5",
+    "@babel/preset-env": "^7.22.5",
+    "@babel/preset-react": "^7.22.5",
+    "@babel/preset-typescript": "^7.22.5",
+    "@reduxjs/toolkit": "^1.9.5",
+    "@types/uuid": "^9.0.4",
+    "@typescript-eslint/eslint-plugin": "^5.60.1",
     "babel-loader": "^9.1.2",
-    "dotenv": "^16.0.3",
+    "css-loader": "^6.8.1",
+    "dotenv": "^16.3.1",
     "dotenv-webpack": "^8.0.1",
-    "eslint": "^8.41.0",
+    "eslint": "^8.43.0",
     "eslint-config-standard-with-typescript": "^34.0.1",
     "eslint-plugin-import": "^2.27.5",
     "eslint-plugin-n": "^15.7.0",
     "eslint-plugin-promise": "^6.1.1",
     "eslint-plugin-react": "^7.32.2",
-    "html-webpack-plugin": "^5.5.1",
+    "html-webpack-plugin": "^5.5.3",
+    "lazy": "^1.0.11",
     "react-dom": "^18.2.0",
-    "typescript": "^5.0.4",
-    "webpack": "^5.83.1",
-    "webpack-cli": "^5.1.1",
-    "webpack-dev-server": "^4.15.0",
+    "react-redux": "^8.1.1",
+    "react-router": "^6.14.1",
+    "react-router-dom": "^6.14.1",
+    "sass": "^1.63.6",
+    "sass-loader": "^13.3.2",
+    "style-loader": "^3.3.3",
+    "tailwindcss": "^3.3.3",
+    "typescript": "^5.1.6",
+    "webpack": "^5.88.1",
+    "webpack-cli": "^5.1.4",
+    "webpack-dev-server": "^4.15.1",
     "webpack-merge": "^5.9.0"
   }
 }
