@@ -1,10 +1,18 @@
+import { BrowserRouter } from 'react-router-dom'
+import { createRoot } from 'react-dom/client'
+import { Provider } from 'react-redux'
 import React from 'react'
-import ReactDOM from 'react-dom'
 import App from './App'
+import { store } from '@apps/store'
 
-ReactDOM.render(
+const container = document.getElementById('root')
+const root = createRoot(container!!)
+root.render(
   <React.StrictMode>
-    <App />
-  </React.StrictMode>,
-  document.getElementById('root')
+    <BrowserRouter>
+      <Provider store={store}>
+        <App />
+      </Provider>
+    </BrowserRouter>
+  </React.StrictMode>
 )
