@@ -5,7 +5,18 @@ const HtmlWebpackPlugin = require('html-webpack-plugin');
 module.exports = {
   entry: path.resolve(__dirname, '../src/index.tsx'),
   resolve: {
-    extensions: ['.tsx', '.ts', '.js']
+    extensions: ['.tsx', '.ts', '.js'],
+    alias: {
+      "@apps": path.resolve(__dirname, "../src/apps"),
+      "@assets": path.resolve(__dirname, "../src/assets"),
+      "@enums": path.resolve(__dirname, "../src/enums"),
+      "@pages": path.resolve(__dirname, "../src/pages"),
+      "@routes": path.resolve(__dirname, "../src/routes"),
+      "@utils": path.resolve(__dirname, "../src/utils"),
+      "@context": path.resolve(__dirname, "../src/context"),
+      "@globalTypes": path.resolve(__dirname, "../src/types"),
+      "@components": path.resolve(__dirname, "../src/components"),
+    },
   },
   plugins: [
     new HtmlWebpackPlugin({
@@ -31,11 +42,11 @@ module.exports = {
             loader: 'babel-loader'
           }
         ]
-      }
-      // {
-      //     test: /\.(?:css|scss)$/,
-      //     use: ['style-loader', 'css-loader', 'sass-loader'],
-      // },
+      },
+      {
+          test: /\.(?:css|scss)$/,
+          use: ['style-loader', 'css-loader', 'sass-loader'],
+      },
       // {
       //     test: /\.(?:ico|gif|png|jpg|jpeg)$/i,
       //     type: 'asset/resource',
