@@ -1,11 +1,9 @@
 module.exports = {
   mode: 'development',
-
-
-  
   devServer: {
     hot: true,
     open: true,
+    historyApiFallback: true,
   },
   devtool: 'cheap-module-source-map',
 }
